# public/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Flowy API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Flowy",
      "url": "https://github.com/GabrielCrackPro/flowy"
    },
    "description": "Welcome to the **Flowy API**. Use it to manage transactions, budgets, savings goals, subscriptions, categories, and shared spaces.\n\nUse the sidebar or search (`Ctrl/Cmd + K`) to find an endpoint. Each operation includes its authentication, parameters, request body, response examples, and common errors.\n\n## Base URL\nProduction requests use `https://flowy-jade.vercel.app`. Local and preview deployments use the same-origin server. JSON endpoints accept and return `application/json` unless noted otherwise.\n\n## Authentication\nSend a Supabase access token with `Authorization: Bearer <access_token>`. Your requests use the authenticated user's active space; personal data is kept in the personal space.\n\n## Errors and limits\nErrors use a consistent `message` and `category` shape. Validation errors can include field-level details in `errors`. If a request is rate-limited, wait for `Retry-After` before retrying."
  },
  "servers": [
    {
      "url": "https://flowy-jade.vercel.app",
      "description": "Production"
    },
    {
      "url": "/",
      "description": "Same-origin — local development and preview deployments"
    }
  ],
...
  "paths": {
...
    "/api/account/password": {
      "post": {
        "operationId": "account.changePassword",
        "summary": "Change password",
        "tags": [
          "Account"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Verifies the current password, then updates it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "Common error response returned when a request cannot be completed. Validation responses may include field-level details in `errors`.",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "category": {
            "type": "string",
            "description": "Broad error category for client-side handling"
          },
          "severity": {
            "type": [
              "string",
              "null"
            ],
            "description": "Severity level, when assigned"
          },
          "statusCode": {
            "type": [
              "integer",
              "null"
            ],
            "description": "HTTP status code"
          },
          "isRetryable": {
            "type": "boolean",
            "description": "Whether retrying may succeed"
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Zod validation issues (message + errors only on 400s)"
          },
          "retryAfter": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Seconds until the rate-limit window resets"
          },
          "retryAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO time when the rate-limit window resets"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "DeleteAccountResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Confirmation message"
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "The user's current password"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "description": "New password (min 8 characters)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is invalid. Check the response's `errors` details and verify the parameters or JSON body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is missing or invalid. Send a valid Supabase access token as `Authorization: Bearer <access_token>`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "The server could not complete the request. Try again later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase access token sent as `Authorization: Bearer <access_token>`."
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal scheduler credential. Not used by client applications."
      }
    }
  }
}